pyproject.toml
[project]
name = "kandown"
version = "0.1.0"
description = "A markdown backed kanban board easy to use"
readme = "README.md"
authors = [
    { name = "Maic Siemering", email = "maic@siemering.tech" }
]
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.0",
    "markdown>=3.6.0",
    "click>=8.1.0",
    "pyyaml>=6.0.3",
    "atomicwrites>=1.4.1",
    "pydantic>=2.11.10",
]

[project.scripts]
kandown = "kandown.cli:main"

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"


[tool.pytest.ini_options]
markers = [
    "e2e: end to end tests",
]

[tools.pytest.ini_options]
addopts = "--tracing retain-on-failure --headed --browser chromium"


[tool.ruff]
line-length = 120

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-playwright>=0.7.1",
    "ruff>=0.13.2",
]
